Begin flatten refactor
The API surface was too large and too complex to understand, and this
made various operations harder. This has now been flattened massively.
This includes removing the singleton pattern from the engine and
ensuring that config objects are just simpler to understand.

Also, the daemon process has been explicitly removed and been replaced
with the "agent" subcommand from the primary orbit CLI.