Add initial failing implementation of leader forwarding
I have no idea why this isn’t working. As it stands, whenever I attempt to make a request through Postman, the request errors out with “Could not get any response” and “There was an error connecting to…”.

As far as I can tell, the connection is getting aborted and stopped before the redirect response can get handled. It is a POST request, but I don’t know if that has anything to do with it. It is interesting though, and in future I think I will need to figure out a way to get this to work. The funny thing is that after looking through the code for `gin-gonic/gin`, it really appears as though it just uses the underlying Golang http handler, which makes it even weirder that it’s not working.

Oh well, time to attempt a reverse proxy.