Fix cluster join not waiting for raft to finish
This was a weird one. Basically, raft hadn’t acquired a leader address by this point on the local node, and so adding the node information to the store state failed even through leader forwarding is implemented (there was no leader, as far as it was aware, that it was able to forward to).