Add more safeguards and checks to rpc.Leader()
This is because our node list has not yet had the chance to get updated by Raft when we first called this, which led to race conditions. This means that we need to add a timeout and wait for this get updated. If we don’t manage to retrieve that data within 20 seconds, then fail the operation as we don’t have the leader address.